+++ src/css/exercises.css
@@ -228,8 +228,35 @@
 }
 
 .exercises-gallery-link {
-  display: block;
-  color: #757575;
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  align-items: center;
+  justify-content: center;
+  padding: 45px 105px;
+  width: 295px;
+  height: 232px;
+  /* color: #757575; */
+
+  border-radius: 12px;
+  border: 1px solid rgba(246, 246, 246, 0.3);
+
+  background-image: linear-gradient(
+      0deg,
+      rgba(16, 16, 16, 0.7) 0%,
+      rgba(16, 16, 16, 0.7) 100%
+    ),
+    url(../img/advice/3-110-mob.webp);
+  /* lightgray -16.825px 0.844px / 121.36% 108.504% */
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+@media (min-width: 768px) {
+  .exercises-gallery-link {
+    width: 313px;
+    height: 250px;
+  }
 }
 
 .exercises-gallery-item-area {
@@ -238,41 +265,46 @@
 }
 
 .exercises-gallery-item-title {
-  position: absolute;
+  color: #f6f6f6;
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 20px;
+
+  /* position: absolute;
   bottom: 0;
   left: 0;
   width: 100%;
 
-  height: 100%;
-  background-color: rgba(16, 16, 16, 0.7);
-  color: rgba(246, 246, 246, 1);
+  height: 100%; */
+  /* background-color: rgba(16, 16, 16, 0.7);
+  color: rgba(246, 246, 246, 1); */
 
-  display: flex;
+  /* display: flex;
   align-items: center;
   justify-content: center;
   flex-direction: column;
 
-  border-radius: 12px;
+  border-radius: 12px; */
 }
 
 .exercises-gallery-item-title:hover,
 .exercises-gallery-item-title:focus {
-  background-color: rgba(16, 16, 16, 0.6);
+  /* background-color: rgba(16, 16, 16, 0.6); */
 }
 
-.exercises-gallery-item-title h2 {
+/* .exercises-gallery-item-title h2 {
   font-weight: 400;
   font-size: 20px;
   line-height: 20px;
-}
+} */
 
-@media screen and (min-width: 768px) {
+/* @media screen and (min-width: 768px) {
   .exercises-gallery-item-title h2 {
     font-weight: 400;
     font-size: 24px;
     line-height: 24px;
   }
-}
+} */
 
 .exercises-gallery-item-description {
   font-weight: 400;
