+++ src/js/excercises.js
@@ -1,21 +1,1 @@
-const exercisesType = document.querySelector('.exercises-type-item-btn');
-const exercisesGalleryType = document.querySelector('.exercises-gallery-list');
-// const typeBtn = []
 
-document.addEventListener("DOMContentLoaded", () => {
-  // alert("Page is loaded!");
-  exercisesType.classList.add('exercises-type-item-btn-focus');
-
-  console.log(exercisesType.);
-});
-
-exercisesType.addEventListener('click', () => {
-  exercisesType.classList.remove('exercises-type-item-btn-focus');
-  // alert("Page is loaded!");
-});
-
-
-
-// function selectTypeExercises() {
-//   console.log('11111111');  
-// };
